app/requirements.txt
streamlit 
PyPDF2 
langchain 
openai
langchain_openai
python-dotenv
langchain_community
chromadb
tiktoken
pypdf
python-dotenv